#84 - Moving TagContext into /components
