+++ src/pages/[lang]/projects/index.astro
@@ -1,11 +1,11 @@
 ---
 import { createSupabaseServerClient } from '@backend/supabaseServerClient';
+import { TagContextProvider } from '@components/TagContext';
 import { getLangFromUrl, getTranslations } from '@i18n';
 import { getMyProfile, listMyInvites } from '@backend/crud';
 import { ProjectsHomeWrapper } from '@apps/dashboard-projects';
 import { listMyProjectsExtended } from '@backend/helpers';
 import BaseLayout from '@layouts/BaseLayout.astro';
-import { TagContextProvider } from '@util/context';
 
 const lang = getLangFromUrl(Astro.url);
