+++ src/util/index.ts
@@ -1,3 +1,2 @@
-export * from './context';
 export * from './export';
 export * from './iiif';
